Prepare for 0.10.1 release
- Update dependencies

- Fix issues that 2.13.7 compiler finds that 3.1.0 compiler does not
flag (numeric widening, unused imports)

+++ core/shared/src/test/scala/doodle/core/CoordinateSpec.scala
@@ -10,22 +10,25 @@ object CoordinateSpec extends Properties("Coordinate properties") {
 
   property("oneHundredPercent evaluates correctly") = forAll { (edge: Int) =>
     val actualEdge = Math.abs(edge)
-    Coordinate.oneHundredPercent.eval(-actualEdge, actualEdge) ?= actualEdge
+    Coordinate.oneHundredPercent.eval(
+      -actualEdge.toDouble,
+      actualEdge.toDouble
+    ) ?= actualEdge.toDouble
   }
 
   property("minusOneHundredPercent evaluates correctly") = forAll {
     (edge: Int) =>
       val actualEdge = Math.abs(edge)
       Coordinate.minusOneHundredPercent.eval(
-        -actualEdge,
-        actualEdge
-      ) ?= -actualEdge
+        -actualEdge.toDouble,
+        actualEdge.toDouble
+      ) ?= -actualEdge.toDouble
   }
 
   property("addition of percent and point is correct") =
     forAllNoShrink(smallNumber, smallNumber) { (percent, point) =>
       val coord = Coordinate.percent(percent) + Coordinate.point(point)
-      val pt = coord.eval(-200, 200)
+      val pt = coord.eval(-200.0, 200.0)
       val expected = point + ((percent / 100.0) * 200)
 
       (expected ~= pt) :| s"percent $percent point $point, expected $expected, actual $pt"
@@ -34,7 +37,7 @@ object CoordinateSpec extends Properties("Coordinate properties") {
   property("subtraction of percent and point is correct") =
     forAllNoShrink(smallNumber, smallNumber) { (percent, point) =>
       val coord = Coordinate.point(point) - Coordinate.percent(percent)
-      val pt = coord.eval(-200, 200)
+      val pt = coord.eval(-200.0, 200.0)
       val expected = point - ((percent / 100.0) * 200)
 
       (expected ~= pt) :| s"percent $percent point $point, expected $expected, actual $pt"
